pyproject.toml
[project]
name = "mem0-mcp"
version = "0.1.0"
description = "MCP server for integrating long term memory into AI agents with Mem0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "vecs>=0.4.5",
    "openai>=1.90.0",
    "python-dotenv>=1.0.1",
    "mem0ai @ git+https://github.com/fenghaitao/mem0.git",
    "litellm @ git+https://github.com/fenghaitao/litellm.git"
]
